\name{mi.mix}
\alias{mi.mix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Multiple imputation from a matrix of probabilities of being MCAR for each missing value.
}
\description{
This function allows imputing data sets with a multiple imputation strategy.
}
\usage{
mi.mix(tab, tab.imp, prob.MCAR, conditions, repbio=NULL, reptech=NULL, nb.iter=3, nknn=15,
weight=1, selec="all", siz=500, ind.comp=1, methodi="mle", q=0.95,
progress.bar=TRUE, details=FALSE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tab}{A data matrix containing numeric and missing values. Each column of this matrix is assumed to correspond to an experimental sample, and each row to an identified peptide.}
  \item{tab.imp}{A matrix where the missing values of \code{tab} have been imputed under the assumption that they are all MCAR. For instance, such a matrix can be obtained from the function \code{\link{impute.slsa}} of this package.}
  \item{prob.MCAR}{A matrix of probabilities that each missing value is MCAR. For instance such a matrix can be obtained from the function \code{\link{prob.mcar.tab}} of this package.}
  \item{conditions}{A vector of factors indicating the biological condition to which each column (experimental sample) belongs.}
  \item{repbio}{A vector of factors indicating the biological replicate to which each column belongs. Default is NULL (no experimental design is considered).}
  \item{reptech}{A vector of factors indicating the technical replicate to which each column belongs. Default is NULL (no experimental design is considered).}
  \item{nb.iter}{The number of iterations used for the multiple imputation method.}
  \item{nknn}{The number of nearest neighbours used in the SLSA algorithm (see \code{\link{impute.slsa}}).}
  \item{selec}{A parameter to select a part of the dataset to find nearest neighbours between rows. This can be useful for big data sets (see \code{\link{impute.slsa}}).}
  \item{siz}{A parameter to select a part of the dataset to perform imputations with the SLSA or the MLE algorithm. This can be useful for big data sets. Note that \code{siz} needs to be inferior to \code{selec}.}
  \item{weight}{The way of weighting in the algorithm (see \code{\link{impute.slsa}}).}
  \item{ind.comp}{If \code{ind.comp=1}, only nearest neighbours without missing values are selected to fit linear models (see \code{\link{impute.slsa}}). Else, they can contain missing values.}
  \item{methodi}{The method used for imputing data. If \code{methodi="mle"} (default), then the MLE algorithm is used (function \code{impute.wrapper.MLE} of the R package imputeLCMD), else the SLSA algorithm is used.}
  \item{q}{A quantile value (see \code{\link{impute.igcda}}).}
  \item{progress.bar}{If \code{TRUE}, a progress bar is displayed.}
  \item{details}{If \code{TRUE}, the function gives a list of three values: \code{imputed.matrix} a matrix with the average of imputed values for each missing value, \code{sd.imputed.matrix} a matrix with the standard deviations of imputed values for each missing value, \code{all.imputed.matrices} an array with all the \code{nb.iter} matrices of imputed values that have been generated.}
}
\details{
At each iteration, a matrix indicating the MCAR values is generated by Bernouilli distributions having parameters given by the matrix \code{prob.MCAR}. The generated MCAR values are next imputed thanks to the matrix \code{tab.imp}. For each row containing MNAR values, the other rows are imputed thanks to the function  \code{\link{impute.igcda}} and, next, the considered row is imputed thanks to either the function \code{\link{impute.slsa}} or the function \code{impute.wrapper.MLE} of the R package imputeLCMD. So, the function \code{\link{impute.igcda}} allows to deform the correlation structure of the dataset in view to be closer to that of the true values, while the function \code{\link{impute.slsa}} (\code{impute.wrapper.MLE}) will impute by taking into account this modified correlation structure.
}
\value{
The input matrix \code{tab} with average imputed values instead of missing values if \code{details=FALSE} (default). If \code{details=TRUE}, a list of three values: \code{imputed.matrix} a matrix with the average of imputed values for each missing value, \code{sd.imputed.matrix} a matrix with the standard deviations of imputed values for each missing value, \code{all.imputed.matrices} an array with all the \code{nb.iter} matrices of imputed values that have been generated.
}
%\references{
%% ~put references to the literature/web site here ~
%}
\author{
Quentin Giai Gianetto <quentin2g@yahoo.fr>
}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%}
\examples{
\donttest{
#Simulating data
res.sim=sim.data(nb.pept=5000,nb.miss=1000);

#Fast imputation of missing values with the impute.rand algorithm
dat.rand=impute.rand(tab=res.sim$dat.obs,conditions=res.sim$condition);

#Estimation of the mixture model
res=estim.mix(tab=res.sim$dat.obs, tab.imp=dat.rand, conditions=res.sim$condition);

#Computing probabilities to be MCAR
born=estim.bound(tab=res.sim$dat.obs,conditions=res.sim$condition);
proba=prob.mcar.tab(born$tab.lower,born$tab.upper,res);

#Multiple imputation strategy with 3 iterations (can be time consuming in function of the data set!)
data.mi=mi.mix(tab=res.sim$dat.obs, tab.imp=dat.rand, prob.MCAR=proba, conditions=
res.sim$conditions, repbio=res.sim$repbio, nb.iter=3);
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{Imputation methods}
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
